+++ gomatic/__init__.py
@@ -1,17 +1,9 @@
-from .go_cd_configurator import (Agent, BuildArtifact, ExecTask,
-                                 FetchArtifactDir, FetchArtifactFile,
-                                 FetchArtifactTask,
-                                 GoCdConfigurator, HostRestClient ,
-                                 Pipeline, PipelineGroup, PipelineMaterial,
-                                  Tab, TestArtifact,
-                                 ScriptExecutorTask,  Artifact)
-
-#from .go_cd_configurator_test import FakeHostRestClient, empty_config_xml
+from .go_cd_configurator import ( GoCdConfigurator, HostRestClient )
 from gomatic.go_cd_configurator import HostRestClient, GoCdConfigurator
 from gomatic.gocd.agents import Agent
 from gomatic.gocd.materials import GitMaterial, PipelineMaterial
 from gomatic.gocd.pipelines import Tab, Job, Pipeline, PipelineGroup
-from gomatic.gocd.tasks import FetchArtifactTask, ExecTask, RakeTask, MavenTask
-from gomatic.gocd.artifacts import FetchArtifactFile, FetchArtifactDir, BuildArtifact, TestArtifact, ArtifactFor
+from gomatic.gocd.tasks import FetchArtifactTask, ExecTask, RakeTask, MavenTask, ScriptExecutorTask
+from gomatic.gocd.artifacts import FetchArtifactFile, FetchArtifactDir, BuildArtifact, TestArtifact, ArtifactFor, Artifact
 from gomatic.fake import FakeHostRestClient, empty_config
