+++ gomatic/gocd/pipelines.py
@@ -10,7 +10,6 @@ from collections import OrderedDict
 
 DEFAULT_LABEL_TEMPLATE = "0.${COUNT}"  # TODO confirm what default really is. I am pretty sure this is mistaken!
 
-
 class Tab(CommonEqualityMixin):
     def __init__(self, name, path):
         self.__name = name
@@ -19,9 +18,17 @@ class Tab(CommonEqualityMixin):
     def __repr__(self):
         return 'Tab("%s", "%s")' % (self.__name, self.__path)
 
+    def to_dict(self, ordered=False):
+        if ordered:
+            result = OrderedDict()
+        else:
+            result = {}
+        result['name'] = self.__name
+        result['path'] = self.__path
+        return result
+
     def append_to(self, element):
         element.append(ET.fromstring('<tab name="%s" path="%s" />' % (self.__name, self.__path)))
-
 
 class Job(CommonEqualityMixin):
     def __init__(self, element):
@@ -42,7 +49,6 @@ class Job(CommonEqualityMixin):
             result['timeout'] = self.timeout
         if self.runs_on_all_agents:
             result['runs_on_all_agents'] = True
-        ####   MUST BE FIXED!!!!!
         result['tasks'] = [t.to_dict(ordered=ordered) for t in self.tasks]
         result['artifacts'] = [a.to_dict(ordered=ordered)
                                for a in self.artifacts]
@@ -277,7 +283,7 @@ class Stage(CommonEqualityMixin):
 
 
 
-@property
+    @property
     def has_manual_approval(self):
         return PossiblyMissingElement(self.element).possibly_missing_child("approval").has_attribute("type", "manual")
 
@@ -338,24 +344,29 @@ class Pipeline(CommonEqualityMixin):
         else:
             result = {}
         result['name'] = self.name
-        result['group'] = group_name
-        if self.has_label_template:
-            result['label_template'] = self.label_template()
-        result['automatic_pipeline_locking'] = \
-            self.has_automatic_pipeline_locking
-        result['cron_timer_spec'] = self.timer if self.has_timer else None
-        if self.has_timer:
-            result['cron_timer_run_only_on_new_material'] = \
-                self.timer_triggers_only_on_changes()
-        result['materials'] = [m.to_dict(ordered=ordered)
-                               for m in self.materials]
-        if self.__template_name:
-            result['template'] = self.__template_name
+
+        # template is pipeline with a bit different set of fields
+        if not self.is_template:
+
+            result['group'] = group_name
+            if self.has_label_template:
+                result['label_template'] = self.label_template
+            result['automatic_pipeline_locking'] = \
+                self.has_automatic_pipeline_locking
+            result['cron_timer_spec'] = self.timer if self.has_timer else None
+            if self.has_timer:
+                result['cron_timer_run_only_on_new_material'] = \
+                    self.timer_triggers_only_on_changes()
+            result['materials'] = [m.to_dict(ordered=ordered)
+                                   for m in self.materials]
+            result['environment_variables'] = self.environment_variables
+            result['encrypted_environment_variables'] = \
+                self.encrypted_environment_variables
+            result['parameters'] = self.parameters
+            if self.__template_name:
+                result['template'] = self.__template_name
+
         result['stages'] = [s.to_dict(ordered=ordered) for s in self.stages]
-        result['environment_variables'] = self.environment_variables
-        result['encrypted_environment_variables'] = \
-            self.encrypted_environment_variables
-        result['parameters'] = self.parameters
 
         return result
